プロダクトコードをappフォルダに移動して、アプリケーション起動用に `run.py` を作成しました。
